requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
SQLAlchemy==2.0.21
Faker==19.6.1
pydantic==2.3.0
pandas==2.2.2
google-genai==0.3.0
